# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  apimodels.LoginUserRequest:
    properties:
      emailOrLogin:
        type: string
      password:
        type: string
    type: object
  apimodels.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  apimodels.SignupUserRequest:
    properties:
      email:
        type: string
      login:
        type: string
      password:
        type: string
    type: object
...
  apimodels.Token:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
info:
  contact: {}
  description: Swagger API for Dreamit-based web labs
  termsOfService: http://swagger.io/terms/
  title: Dreamit Swagger API
  version: "2.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get users collection
      parameters:
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get users collection
      tags:
      - user
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user
      tags:
      - user
  /users/login:
    post:
      consumes:
      - application/json
      description: Logging user in by retrieving his data from the database
      parameters:
      - description: User log in data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apimodels.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apimodels.Token'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      summary: Logging user in
      tags:
      - user
  /users/logout:
    post:
      description: Logging user in out
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      summary: Logging user out
      tags:
      - user
  /users/signup:
    post:
      consumes:
      - application/json
      description: Signing user up by adding him to the database
      parameters:
      - description: User sign up data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apimodels.SignupUserRequest'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      summary: Signing user up
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"